building a boss enemy

+++ src/com/company/BasicEnemy.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class BasicEnemy extends GameObject {
 
-    public BasicEnemy(int x, int y, ID id, Handler handler) {
+    public BasicEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id, handler);
         velX = 5;
         velY = 5;
@@ -25,12 +25,12 @@ public class BasicEnemy extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int)x,(int)y,16,16);
 
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x,y,16,16);
+        return new Rectangle((int)x,(int)y,16,16);
     }
 }
