+++ src/com/company/HUD.java
@@ -4,15 +4,15 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
-    private int greenValue = 255;
+    public static float HEALTH = 100.0f;
+    private float greenValue = 255.0f;
     private int score = 0;
     private int level = 1;
 
     public void tick(){
 
-        HEALTH = Game.clamp(HEALTH, 0 , 100);
-        greenValue = Game.clamp(greenValue, 0, 255);
+        HEALTH = Game.clamp((float)HEALTH, 0f , 100.0f);
+        greenValue = Game.clamp((float) greenValue, 0f, 255.0f);
         greenValue = HEALTH * 2;
 
         score++;
@@ -22,8 +22,8 @@ public class HUD {
     public void render(Graphics g){
         g.setColor(Color.gray);
         g.fillRect(15,15, 200, 32);
-        g.setColor(new Color(75,greenValue, 0));
-        g.fillRect(15,15, HEALTH * 2, 32);
+        g.setColor(new Color(75,(int)greenValue, 0));
+        g.fillRect(15,15, (int)HEALTH * 2, 32);
         g.setColor(Color.white);
         g.drawRect(15,15, 200, 32);
